Make prompt more compact
Most segments print with a space on either side of them that can't be
removed by layouts. In order to support compact layouts, those spaces
can be accounted for in the separator instead of being hard-coded. I
didn't change the host segment for now, because that segment seems to
handle separators a bit differently