Git segment: Support submodules outside of a git repo
Submodules use a `.git` file instead of a `.git` directory. In most cases this
is fine since one of the parent directories will have a `.git` directory. In
some cases (using a bare repo to version control dotfiles, for example), a
submodule may not have any parent directory with a .git directory.